+++ tracker/tracker/src/main/app/canvas.ts
@@ -5,12 +5,15 @@ import Message, { CanvasNode } from './messages.gen.js'
 interface CanvasSnapshot {
   images: { data: string; id: number }[]
   createdAt: number
+  paused: boolean
+  dummy: HTMLCanvasElement
 }
 
 interface Options {
   fps: number
   quality: 'low' | 'medium' | 'high'
   isDebug?: boolean
+  fixedScaling?: boolean
 }
 
 class CanvasRecorder {
@@ -27,19 +30,19 @@ class CanvasRecorder {
 
   startTracking() {
     setTimeout(() => {
-      this.app.nodes.scanTree(this.handleCanvasEl)
+      this.app.nodes.scanTree(this.captureCanvas)
       this.app.nodes.attachNodeCallback((node: Node): void => {
-        this.handleCanvasEl(node)
+        this.captureCanvas(node)
       })
     }, 500)
   }
 
   restartTracking = () => {
     this.clear()
-    this.app.nodes.scanTree(this.handleCanvasEl)
+    this.app.nodes.scanTree(this.captureCanvas)
   }
 
-  handleCanvasEl = (node: Node) => {
+  captureCanvas = (node: Node) => {
     const id = this.app.nodes.getID(node)
     if (!id || !hasTag(node, 'canvas')) {
       return
@@ -49,10 +52,41 @@ class CanvasRecorder {
     if (isIgnored || !hasTag(node, 'canvas') || this.snapshots[id]) {
       return
     }
+
+    const observer = new IntersectionObserver((entries) => {
+      entries.forEach((entry) => {
+        if (entry.isIntersecting) {
+          if (entry.target) {
+            if (this.snapshots[id] && this.snapshots[id].createdAt) {
+              this.snapshots[id].paused = false
+            } else {
+              this.recordCanvas(entry.target, id)
+            }
+            /**
+             * We can switch this to start observing when element is in the view
+             * but otherwise right now we're just pausing when it's not
+             * just to save some bandwidth and space on backend
+             * */
+            // observer.unobserve(entry.target)
+          } else {
+            if (this.snapshots[id]) {
+              this.snapshots[id].paused = true
+            }
+          }
+        }
+      })
+    })
+
+    observer.observe(node)
+  }
+
+  recordCanvas = (node: Node, id: number) => {
     const ts = this.app.timestamp()
     this.snapshots[id] = {
       images: [],
       createdAt: ts,
+      paused: false,
+      dummy: document.createElement('canvas'),
     }
     const canvasMsg = CanvasNode(id.toString(), ts)
     this.app.send(canvasMsg as Message)
@@ -63,11 +97,18 @@ class CanvasRecorder {
         this.app.debug.log('Canvas element not in sync')
         clearInterval(int)
       } else {
-        const snapshot = captureSnapshot(canvas, this.options.quality)
-        this.snapshots[id].images.push({ id: this.app.timestamp(), data: snapshot })
-        if (this.snapshots[id].images.length > 9) {
-          this.sendSnaps(this.snapshots[id].images, id, this.snapshots[id].createdAt)
-          this.snapshots[id].images = []
+        if (!this.snapshots[id].paused) {
+          const snapshot = captureSnapshot(
+            canvas,
+            this.options.quality,
+            this.snapshots[id].dummy,
+            this.options.fixedScaling,
+          )
+          this.snapshots[id].images.push({ id: this.app.timestamp(), data: snapshot })
+          if (this.snapshots[id].images.length > 9) {
+            this.sendSnaps(this.snapshots[id].images, id, this.snapshots[id].createdAt)
+            this.snapshots[id].images = []
+          }
         }
       }
     }, this.interval)
@@ -110,18 +151,36 @@ class CanvasRecorder {
 }
 
 const qualityInt = {
-  low: 0.33,
+  low: 0.35,
   medium: 0.55,
   high: 0.8,
 }
 
-function captureSnapshot(canvas: HTMLCanvasElement, quality: 'low' | 'medium' | 'high' = 'medium') {
+function captureSnapshot(
+  canvas: HTMLCanvasElement,
+  quality: 'low' | 'medium' | 'high' = 'medium',
+  dummy: HTMLCanvasElement,
+  fixedScaling = false,
+) {
   const imageFormat = 'image/jpeg' // or /png'
-  return canvas.toDataURL(imageFormat, qualityInt[quality])
+  if (fixedScaling) {
+    const canvasScaleRatio = window.devicePixelRatio || 1
+    dummy.width = canvas.width / canvasScaleRatio
+    dummy.height = canvas.height / canvasScaleRatio
+    const ctx = dummy.getContext('2d')
+    if (!ctx) {
+      return ''
+    }
+    ctx.drawImage(canvas, 0, 0, dummy.width, dummy.height)
+    return dummy.toDataURL(imageFormat, qualityInt[quality])
+  } else {
+    return canvas.toDataURL(imageFormat, qualityInt[quality])
+  }
 }
 
 function dataUrlToBlob(dataUrl: string): [Blob, Uint8Array] | null {
   const [header, base64] = dataUrl.split(',')
+  if (!header || !base64) return null
   const encParts = header.match(/:(.*?);/)
   if (!encParts) return null
   const mime = encParts[1]